code; Int

+++ A.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    double M = 0;
+    int M = 0;
     std::cout << "Hello World!\n"
 
               << "Max"
